seller page controller edit[images]

+++ app/Http/Controllers/Dashboard/SellerController.php
@@ -80,8 +80,8 @@ class SellerController extends Controller
         $this->validate($request,[
             'desc'                     => 'required|min:10|max:255',
             'desc_ar'                  => 'required|min:10|max:255|',
-            'bg'                       => 'required|image|mimes:jpeg,jpg,png,gif,svg|max:2048',
-            'avatar'                   => 'required|image|mimes:jpeg,jpg,png,gif,svg|max:2048',
+            'bg'                       => 'nullable|image|mimes:jpeg,jpg,png,gif,svg|max:2048',
+            'avatar'                   => 'nullable|image|mimes:jpeg,jpg,png,gif,svg|max:2048',
             'governorate_id'           => 'required|integer|exists:governorates,id',
             'city_id'                  => 'required|integer|exists:cities,id',
             'lat'                      => 'required',
@@ -100,16 +100,19 @@ class SellerController extends Controller
         $seller->street             = $request->street;
         $seller->facebook           = $request->facebook;
         $seller->instagram          = $request->instagram;
+
+
         // Store avatar
-    if($request->file('avatar')){
+        if($request->file('avatar')){
         $Image_id = self::fileAvatar($request->file('avatar'),$seller->avatar);
         $seller->avatar = $Image_id;
-    }
+         }
         // Store avatar
-    if($request->file('bg')){
-        $Image_id = self::fileBackground($request->file('bg'),$seller->bg);
-        $seller->bg = $Image_id;
-    }
+        if($request->file('bg')){
+
+            $Image_id = self::fileBackground($request->file('bg'),$seller->bg);
+            $seller->bg = $Image_id;
+        }
         $seller->save();
 
         session()->flash('updated',__("Changes has been Updated successfully"));
@@ -118,46 +121,69 @@ class SellerController extends Controller
     // store avatar function -------------------------------------------
     public static function fileAvatar($file,$id)
     {
+        if(Image::find($id))
+        {
         $Image = Image::find($id);
         $extension = $file->getClientOriginalExtension();
         $fileName = time() . rand(11111, 99999) . '.' . $extension;
         $destinationPath = public_path() . $Image->base;
         $file->move($destinationPath, $fileName);
         //Delete Old image
-
         try {
-            $file_old = $destinationPath.$Image->id;
+            $file_old = $destinationPath . $Image->name;
             unlink($file_old);
             $Image->delete();
         } catch (Exception $e) {
             echo 'Caught exception: ',  $e->getMessage(), "\n";
         }
         //Update new image
-        $Image = Image::create(['name'=> $fileName, 'base' =>  '/img/avatar/']);
+        $Image = Image::create(['name'=> $fileName, 'base' =>  $Image->base]);
         return $Image->id;
+
+        } else {
+        $extension = $file->getClientOriginalExtension();
+        $fileName = time() . rand(11111, 99999) . '.' . $extension;
+        $destinationPath = public_path() . '\img\avatar\\';
+        $file->move($destinationPath, $fileName);
+        $Image = Image::create(['name'=> $fileName, 'base' =>  '\img\avatar\\']);
+        return $Image->id;    
+        }
+
     }
    // store background function -------------------------------------------
-    public static function fileBackground($file,$id)
-    {
+   public static function fileBackground($file,$id)
+   {
+       if(Image::find($id))
+       {
         $Image = Image::find($id);
         $extension = $file->getClientOriginalExtension();
         $fileName = time() . rand(11111, 99999) . '.' . $extension;
         $destinationPath = public_path() . $Image->base;
         $file->move($destinationPath, $fileName);
+  
         //Delete Old image
-
         try {
-            $file_old = $destinationPath.$Image->id;
+            $file_old = $destinationPath . $Image->name;
             unlink($file_old);
             $Image->delete();
         } catch (Exception $e) {
             echo 'Caught exception: ',  $e->getMessage(), "\n";
         }
         //Update new image
-        $Image = Image::create(['name'=> $fileName, 'base' =>  '/img/background/']);
+        $Image = Image::create(['name'=> $fileName, 'base' =>  $Image->base]);
         return $Image->id;
-    }
 
+       } else {
+
+        $extension = $file->getClientOriginalExtension();
+        $fileName = time() . rand(11111, 99999) . '.' . $extension;
+        $destinationPath = public_path() . '\img\background\\';
+        $file->move($destinationPath, $fileName);
+        $Image = Image::create(['name'=> $fileName, 'base' =>  '\img\background\\']);
+        return $Image->id;
+       }
+
+   }
     /**
      * Remove the specified resource from storage.
      *
